feat: Refactor dataset, featuremodel, explore, hubfile, public
