+++ app/modules/auth/models.py
@@ -1,11 +1,9 @@
-import os
 from datetime import datetime, timezone
 
 from flask_login import UserMixin
 from werkzeug.security import generate_password_hash, check_password_hash
 
 from app import db
-from core.configuration.configuration import uploads_folder_name
 
 
 class User(db.Model, UserMixin):
@@ -32,14 +30,6 @@ class User(db.Model, UserMixin):
     def check_password(self, password):
         return check_password_hash(self.password, password)
 
-    def temp_folder(self):
-        return os.path.join(uploads_folder_name(), "temp", str(self.id))
-
-    def save(self):
-        if not self.id:
-            db.session.add(self)
-        db.session.commit()
-
-    def delete(self):
-        db.session.delete(self)
-        db.session.commit()
+    def temp_folder(self) -> str:
+        from app.modules.auth.services import AuthenticationService
+        return AuthenticationService().temp_folder_by_user(self)
