+++ app/modules/featuremodel/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+
+featuremodel_bp = Blueprint('featuremodel', __name__, template_folder='templates')
