+++ app/modules/featuremodel/forms.py
@@ -1,0 +1,6 @@
+from flask_wtf import FlaskForm
+from wtforms import SubmitField
+
+
+class FeaturemodelForm(FlaskForm):
+    submit = SubmitField('Save featuremodel')
