+++ app/modules/featuremodel/seeders.py
@@ -1,0 +1,12 @@
+from core.seeders.BaseSeeder import BaseSeeder
+
+
+class FeaturemodelSeeder(BaseSeeder):
+
+    def run(self):
+
+        data = [
+            # Create any Model object you want to make seed
+        ]
+
+        self.seed(data)
