+++ app/modules/featuremodel/templates/featuremodel/index.html
@@ -1,0 +1,7 @@
+{% extends "base_template.html" %}
+
+{% block title %}View featuremodel{% endblock %}
+
+{% block content %}
+
+{% endblock %}
