+++ app/modules/featuremodel/tests/test_unit.py
@@ -1,0 +1,24 @@
+import pytest
+
+
+@pytest.fixture(scope='module')
+def test_client(test_client):
+    """
+    Extends the test_client fixture to add additional specific data for module testing.
+    """
+    with test_client.application.app_context():
+        # Add HERE new elements to the database that you want to exist in the test context.
+        # DO NOT FORGET to use db.session.add(<element>) and db.session.commit() to save the data.
+        pass
+
+    yield test_client
+
+
+def test_sample_assertion(test_client):
+    """
+    Sample test to verify that the test framework and environment are working correctly.
+    It does not communicate with the Flask application; it only performs a simple assertion to
+    confirm that the tests in this module can be executed.
+    """
+    greeting = "Hello, World!"
+    assert greeting == "Hello, World!", "The greeting does not coincide with 'Hello, World!'"
