+++ app/modules/hubfile/repositories.py
@@ -1,6 +1,7 @@
 from sqlalchemy import func
 from app.modules.auth.models import User
-from app.modules.dataset.models import DataSet, FeatureModel
+from app.modules.dataset.models import DataSet
+from app.modules.featuremodel.models import FeatureModel
 from app.modules.hubfile.models import Hubfile, HubfileDownloadRecord, HubfileViewRecord
 from core.repositories.BaseRepository import BaseRepository
 from app import db
@@ -24,19 +25,19 @@ class HubfileRepository(BaseRepository):
         return db.session.query(DataSet).join(FeatureModel).join(Hubfile).filter(Hubfile.id == hubfile.id).first()
 
 
-class HubfileDownloadRecordRepository(BaseRepository):
-    def __init__(self):
-        super().__init__(HubfileDownloadRecord)
-
-    def total_feature_model_downloads(self) -> int:
-        max_id = self.model.query.with_entities(func.max(self.model.id)).scalar()
-        return max_id if max_id is not None else 0
-
-
 class HubfileViewRecordRepository(BaseRepository):
     def __init__(self):
         super().__init__(HubfileViewRecord)
 
-    def total_feature_model_views(self) -> int:
+    def total_hubfile_views(self) -> int:
+        max_id = self.model.query.with_entities(func.max(self.model.id)).scalar()
+        return max_id if max_id is not None else 0
+
+
+class HubfileDownloadRecordRepository(BaseRepository):
+    def __init__(self):
+        super().__init__(HubfileDownloadRecord)
+
+    def total_hubfile_downloads(self) -> int:
         max_id = self.model.query.with_entities(func.max(self.model.id)).scalar()
         return max_id if max_id is not None else 0
