+++ app/modules/public/routes.py
@@ -2,8 +2,9 @@ import logging
 
 from flask import render_template
 
+from app.modules.featuremodel.services import FeatureModelService
 from app.modules.public import public_bp
-from app.modules.dataset.services import DataSetService, FeatureModelService
+from app.modules.dataset.services import DataSetService
 
 logger = logging.getLogger(__name__)
 
@@ -20,11 +21,11 @@ def index():
 
     # Statistics: total downloads
     total_dataset_downloads = dataset_service.total_dataset_downloads()
-    total_feature_model_downloads = dataset_service.total_feature_model_downloads()
+    total_feature_model_downloads = feature_model_service.total_feature_model_downloads()
 
     # Statistics: total views
     total_dataset_views = dataset_service.total_dataset_views()
-    total_feature_model_views = dataset_service.total_feature_model_views()
+    total_feature_model_views = feature_model_service.total_feature_model_views()
 
     return render_template(
         "public/index.html",
